feat: change pages with ctrl-tab
Use the common browser interaction ctrl-tab to switch pages forward and alt-tab to switch pages back.

Implementation-wise, we have to break a little layering by using the displayPages list of guids that really only the tab bar is supposed to care about... but we have to call ChangePage in the main screen, so I'm not sure how else to do it.

The other change of note is forcing a focus change when you switch tabs; this is to get any active dropdowns (i.e. the GoodsIcon dropdowns for switching recipes or switching modules) to disappear if you hit control-tab while one is open, since the mouse isn't otherwise moving and clicking like it would be if you switched tabs with the mouse.