fix: destroy textures on the main thread
SDL2 is not, generally speaking, threadsafe especially in video APIs and even more so in video APIs that access some shared resource like render driver work queues. Textures, and SDL_DestroyTexture, are one such resource.

However, we've been calling SDL_DestroyTexture in the finalizer of our wrapper TextureHandle. The CLR is generally happy to do things like run finalizers in a GC thread, and therefore we're calling SDL_DestroyTexture from a non-main thread.

We can (slightly) abuse the SDL event loop to fix this by pushing the actual SDL_DestroyTexture call through a user event via the UI base class's static method, which will definitely fix this issue.

It's hard to prove that this is or isn't the cause of these problems, but thread traces where the main thread is running an SDL texture operation while the GC thread is calling SDL_DestroyTexture have popped up a couple times in crash repros, so let's try avoiding it.